part4 finished

+++ part4/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then(blogs => {
+  Blog.find({}).populate('user', { username: 1, name: 1, id: 1})
+  .then(blogs => {
     response.json(blogs)
   })
 })
@@ -19,28 +22,40 @@ blogsRouter.get('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-blogsRouter.post('/', (request, response, next) => {
+
+blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const token = request.token
+  const id = request.user
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
-    url: body.url
+    url: body.url,
+    user: user._id
   })
 
-  blog.save()
-    .then(savedBlog => {
-      response.status(201).json(savedBlog)
-    })
-    .catch(error => next(error))
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then(() => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const user = request.user
+
+  const blog= await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).json({ message: 'blog deleted' })
+  } else {
+    response.status(401).json({ error: 'user not authorized to delete this blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
